+++ auth0/provider_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 
-	"gopkg.in/lallij/auth0.v1/management"
+	"gopkg.in/lallij/auth0.v9/management"
 )
 
 func Auth0() (*management.Management, error) {