+++ auth0/resource_auth0_branding.go
@@ -6,7 +6,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 
-	"gopkg.in/lallij/auth0.v1/management"
+	"gopkg.in/lallij/auth0.v9/management"
 )
 
 func newBranding() *schema.Resource {