+++ auth0/resource_auth0_role.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 
-	"gopkg.in/lallij/auth0.v1"
-	"gopkg.in/lallij/auth0.v1/management"
+	"gopkg.in/lallij/auth0.v9"
+	"gopkg.in/lallij/auth0.v9/management"
 )
 
 func newRole() *schema.Resource {