+++ auth0/resource_auth0_user.go
@@ -10,8 +10,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/structure"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 
-	"gopkg.in/lallij/auth0.v1"
-	"gopkg.in/lallij/auth0.v1/management"
+	"gopkg.in/lallij/auth0.v9"
+	"gopkg.in/lallij/auth0.v9/management"
 )
 
 func newUser() *schema.Resource {