+++ auth0/structure_auth0_tenant.go
@@ -1,6 +1,6 @@
 package auth0
 
-import "gopkg.in/lallij/auth0.v1/management"
+import "gopkg.in/lallij/auth0.v9/management"
 
 func flattenTenantChangePassword(changePassword *management.TenantChangePassword) []interface{} {
 	m := make(map[string]interface{})